+++ lib/screens/reader/main_reader.dart
@@ -4,7 +4,7 @@ import 'package:booketlist/screens/profile.dart';
 import 'package:booketlist/screens/reader/wishlist.dart';
 import 'package:flutter/material.dart';
 import 'package:booketlist/screens/reader/list_buku.dart';
-import 'package:booketlist/screens/reader/homeReader.dart';
+import 'package:booketlist/screens/reader/home_reader.dart';
 import 'package:curved_navigation_bar/curved_navigation_bar.dart';
 
 class MainReader extends StatefulWidget {
